+++ crates/vmi-utils/src/injector/os/windows/kernel_mode.rs
@@ -221,11 +221,8 @@ where
         vmi: &VmiContext<WindowsOs<Driver>>,
     ) -> Result<VmiEventResponse<Amd64>, VmiError> {
         match self.bpm.get_by_event(vmi.event(), ()) {
-            Some(breakpoints) => {
-                // Breakpoints can have multiple tags, but we have set only one
-                // tag for each breakpoint.
-                let first_breakpoint = breakpoints.into_iter().next().expect("breakpoint");
-                debug_assert_eq!(first_breakpoint.tag(), "SeAccessCheck");
+            Some(breakpoint) => {
+                debug_assert_eq!(breakpoint.tag(), "SeAccessCheck");
             }
             None => {
                 if BreakpointController::is_breakpoint(vmi, vmi.event())? {